+++ lib/src/scrobblenaut.dart
@@ -7,6 +7,7 @@ import 'package:meta/meta.dart';
 import 'package:scrobblenaut/src/core/lastfm.dart';
 import 'package:scrobblenaut/src/methods/album_methods.dart';
 import 'package:scrobblenaut/src/methods/artist_methods.dart';
+import 'package:scrobblenaut/src/methods/chart_methods.dart';
 
 /// This connects all the various methods [LastFM] can provide.
 ///
@@ -16,10 +17,12 @@ class Scrobblenaut {
 
   AlbumMethods _albumMethods;
   ArtistMethods _artistMethods;
+  ChartMethods _chartMethods;
 
   Scrobblenaut._(this._api) {
     _albumMethods = AlbumMethods(_api);
     _artistMethods = ArtistMethods(_api);
+    _chartMethods = ChartMethods(_api);
   }
 
   /// It creates a Scrobblenaut Session using a LastFM object.
@@ -33,4 +36,7 @@ class Scrobblenaut {
 
   /// Use this to use [Artist]'s methods.
   ArtistMethods get artist => _artistMethods;
+
+  /// Use this to use [Chart]'s methods.
+  ChartMethods get chart => _chartMethods;
 }